Fix: Split PatchBuilder content test file
